Added snow effects to mountains.
git-svn-id: svn://svn.code.sf.net/p/annchienta/code/trunk@668 6c58e83f-43c0-4801-a1bd-44083660f6ae

+++ Annchienta/fall_of_imiryn/src/battle/Combatant.py
@@ -68,7 +68,7 @@ class BaseCombatant:
         elementalElements = xmlElement.getElementsByTagName("elemental")
         if len(elementalElements):
             for k in elementalElements[0].attributes.keys():
-                self.primaryElemental[k] = int(elementalElements[0].attributes[k].value)
+                self.primaryElemental[k] = float(elementalElements[0].attributes[k].value)
     
         # Generate derived stats
         self.generateDerivedStats()
